Implement ClearAll command
Currently there are only methods to delete one specific students, scores
and graded component respectively and it is very inefficient if we want
to delete all the existing data.

As a step to make the codes more efficient, let's implement a command
so that we can clear the existing data efficiently.