Extend the functionality of SortStudentCommand
Currently SortStudentCommand can only sort the name in alphabetical
order in a decreasing order. As a result, its functionality is very
limited when we need to sort based on other attributes. Let's extend
the functionality by allowing it to sort based on different attributes
and can choose whether to reverse. In addition, let's sort only the
displayed students instead of all students.